refact: side bar slider
