+++ src/views/home-page/components/sidebar-slider/sidebar-slider.module.scss
@@ -1,0 +1,44 @@
+.tachi-sidebar-drag-slide-container {
+    display: flex;
+    &:hover {
+        .tachi-sidebar-drag-slide {
+            border-color: var(--hover-border-color);
+        }
+    }
+}
+
+.tachi-sidebar-drag-slide {
+    width: 1rem;
+    background-color: transparent;
+    cursor: ew-resize !important;
+
+    display: flex;
+    flex-direction: column;
+    .tachi-sidebar-drag-navbar {
+        height: calc(3.6rem);
+        background-color: var(--navbar-bg-color);
+    }
+    .tachi-sidebar-drag-content {
+        flex-grow: 1;
+        background-color: var(--default-bg-color);
+    }
+}
+
+.tachi-sidebar-ldrag-slide {
+    border-right: 0.1rem solid var(--default-border-color);
+}
+
+.tachi-sidebar-rdrag-slide {
+    border-left: 0.1rem solid var(--default-border-color);
+}
+
+.tachi-sidebar-mask {
+    width: 100vw;
+    height: 100vh;
+    background-color: transparent;
+    cursor: ew-resize;
+    
+    position: absolute;
+    top: 0;
+    left: 0;
+}
